reorganization of test folders + swagger made available via localhost + readme.md
