+++ src/application/paymentService.js
@@ -10,7 +10,6 @@ class PaymentService {
         try {
             return await this.paymentUseCase.processPayment(data);
         } catch (error) {
-            console.log('Error in PaymentService:', error);
             throw error; // Rethrow Mercado Pago API errors
         }
     }
@@ -19,7 +18,6 @@ class PaymentService {
         try {
             return await this.paymentUseCase.getPayments(criteria, sort, external_reference);
         } catch (error) {
-            console.log('Error in PaymentService:', error);
             throw error; // Rethrow Mercado Pago API errors
         }
     }
