+++ src/infra/apis/mercadoPagoApi.test.js
@@ -1,5 +1,5 @@
 const {MercadoPagoConfig, Payment} = require('mercadopago');
-const mercadoPagoApi = require('../mercadoPagoApi');
+const mercadoPagoApi = require('./mercadoPagoApi');
 
 jest.mock('mercadopago');
