+++ src/interfaces/swagger/api_doc.yaml
@@ -15,7 +15,7 @@ externalDocs:
   description: Find out more about Swagger
   url: http://swagger.io
 servers:
-  - url: https://localhost:3000/payment/api-docs
+  - url: https://localhost:4004/payment/api-docs
 tags:
   - name: payment
     description: Everything about your payments
